A new hope I guess

+++ src/main/java/slimeknights/mantle/fluid/tooltip/AbstractFluidTooltipProvider.java
@@ -105,18 +105,18 @@ public abstract class AbstractFluidTooltipProvider extends GenericDataProvider {
     private final ImmutableList.Builder<FluidUnit> units = ImmutableList.builder();
 
     /** Adds a unit with a full translation key */
-    public FluidUnitListBuilder addUnitRaw(String key, int amount) {
+    public FluidUnitListBuilder addUnitRaw(String key, long amount) {
       units.add(new FluidUnit(key, amount));
       return this;
     }
 
     /** Adds a unit local to the current mod */
-    public FluidUnitListBuilder addUnit(String key, int amount) {
+    public FluidUnitListBuilder addUnit(String key, long amount) {
       return addUnitRaw(Util.makeDescriptionId("gui", id("fluid." + key)), amount);
     }
 
     /** Adds a unit local to the given mod */
-    public FluidUnitListBuilder addUnit(String key, String domain, int amount) {
+    public FluidUnitListBuilder addUnit(String key, String domain, long amount) {
       return addUnitRaw(Util.makeDescriptionId("gui", new ResourceLocation(domain, "fluid." + key)), amount);
     }
